DjangoAjax/requirements.txt
torchvision==0.7.0
numpy==1.19.2
Django==3.1.2
torch==1.6.0
opencv_python==4.4.0.44
scipy==1.5.0
Pillow==8.0.1
